Add dataframe_regression fixture
