+++ tests/test_dataframe_regression/test_arrays_of_same_size.csv
@@ -1,0 +1,2 @@
+,hello,world
+0,0,0
