+++ tests/test_dataframe_regression/test_arrays_with_different_sizes.csv
@@ -1,0 +1,12 @@
+,data1
+0,1
+1,1
+2,1
+3,1
+4,1
+5,1
+6,1
+7,1
+8,1
+9,1
+10,1
