+++ tests/test_dataframe_regression/test_bool_array.csv
@@ -1,0 +1,4 @@
+,data1
+0,False
+1,False
+2,False
