+++ tests/test_dataframe_regression/test_different_data_types.csv
@@ -1,0 +1,11 @@
+,data1
+0,1
+1,1
+2,1
+3,1
+4,1
+5,1
+6,1
+7,1
+8,1
+9,1
