+++ tests/test_dataframe_regression/test_number_formats.csv
@@ -1,0 +1,4 @@
+,data1
+0,1.2345677999999999e+50
+1,1.2345678e-50
+2,0
